refactor(ui): split workflows._index.tsx god object into modules
