+++ scripts/.god-object-baseline.txt
@@ -3,9 +3,6 @@
 # Doctrine: CLAUDE.md 'Module Size Discipline' — split by responsibility toward <600 LOC.
 1106 ui/src/routes/instances.$id.tsx
 1065 sandbox-runtime/src/session_auth.rs
-922 ui/src/routes/workflows._index.tsx
-744 sandbox-runtime/src/scoped_session_auth.rs
-838 sandbox-runtime/src/reaper.rs
 738 ui/src/routes/sandboxes.$id.tsx
 701 ui/src/lib/stores/chatSessions.ts
 653 ai-agent-instance-blueprint-lib/src/auto_provision.rs
