+++ ui/src/routes/workflows._index.tsx
@@ -1,4 +1,4 @@
-import { type ReactNode, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { Link } from 'react-router';
 import { toast } from 'sonner';
 import { useStore } from '@nanostores/react';
@@ -13,9 +13,7 @@ import { getAddresses, useSubmitJob } from '@tangle-network/blueprint-ui';
 import { encodeJobArgs } from '@tangle-network/blueprint-ui';
 import { getJobById } from '@tangle-network/blueprint-ui';
 import { JOB_IDS, PRICING_TIERS } from '~/lib/types/sandbox';
-import { cn } from '@tangle-network/blueprint-ui';
-import { type Address } from 'viem';
-import { isContractDeployed, type SandboxAddresses } from '~/lib/contracts/chains';
+import { type SandboxAddresses } from '~/lib/contracts/chains';
 import { INSTANCE_OPERATOR_API_URL, OPERATOR_API_URL } from '~/lib/config';
 import { sandboxListStore } from '~/lib/stores/sandboxes';
 import { instanceListStore } from '~/lib/stores/instances';
@@ -27,143 +25,31 @@ import {
   type PendingWorkflowCreation,
 } from '~/lib/stores/pendingWorkflows';
 import {
-  buildWorkflowDetailPath,
   getWorkflowBlueprintIdForScope,
   resolveWorkflowTargetLabelFromValues,
-  type WorkflowBlueprintId,
-  type WorkflowScope,
 } from '~/lib/workflows';
 import {
-  ConsoleChip,
   ConsoleMetricStrip,
   ConsolePage,
   ConsoleSection,
   EmptyConsoleState,
   type ConsoleMetric,
 } from '~/components/console/ConsolePrimitives';
+import { WorkflowActionButton } from '~/components/workflows-list/WorkflowActionButton';
+import { WorkflowTable } from '~/components/workflows-list/WorkflowTable';
 import {
-  IdentityMark,
-  getBlueprintIdentity,
-} from '~/components/shared/VisualIdentity';
+  getOperatorLabel,
+  getWorkflowContractAddressForScope,
+  getWorkflowIdentityKey,
+  getWorkflowSortTimestamp,
+} from '~/components/workflows-list/helpers';
+import {
+  type RemoteWorkflowRecord,
+  type WorkflowRecord,
+} from '~/components/workflows-list/types';
 
 const WORKFLOW_VISIBILITY_POLL_INTERVAL_MS = 3_000;
 const WORKFLOW_VISIBILITY_TIMEOUT_MS = 120_000;
-
-type RemoteWorkflowRecord = {
-  kind: 'remote';
-  id: bigint;
-  scope: WorkflowScope;
-  blueprintId: WorkflowBlueprintId;
-  data: WorkflowOperatorSummary;
-  targetLabel: string;
-  kindLabel: string;
-};
-
-type PendingWorkflowRecord = {
-  kind: 'pending';
-  id: bigint;
-  scope: WorkflowScope;
-  blueprintId: WorkflowBlueprintId;
-  pending: PendingWorkflowCreation;
-  targetLabel: string;
-  kindLabel: string;
-};
-
-type WorkflowRecord = RemoteWorkflowRecord | PendingWorkflowRecord;
-
-function getWorkflowStatusPresentation(workflow: WorkflowOperatorSummary) {
-  if (!workflow.runnable) {
-    return {
-      label: 'Not Runnable',
-      variant: 'stopped' as const,
-      detail: workflow.targetStatus === 'missing'
-        ? 'Target is no longer available'
-        : 'Workflow is currently blocked',
-    };
-  }
-
-  if (workflow.active) {
-    return {
-      label: 'Active',
-      variant: 'running' as const,
-      detail: 'Ready to execute on schedule',
-    };
-  }
-
-  return {
-    label: 'Inactive',
-    variant: 'secondary' as const,
-    detail: 'Disabled until re-enabled',
-  };
-}
-
-function getPendingWorkflowStatusPresentation(pending: PendingWorkflowCreation) {
-  switch (pending.status) {
-    case 'awaiting-auth':
-      return {
-        label: 'Submitted',
-        variant: 'secondary' as const,
-        detail: pending.statusMessage || 'Connect to the operator to verify that the workflow is visible.',
-      };
-    case 'timed-out':
-      return {
-        label: 'Still Processing',
-        variant: 'secondary' as const,
-        detail: pending.statusMessage || 'Creation is taking longer than expected. Check status to look again.',
-      };
-    case 'processing':
-    default:
-      return {
-        label: 'Processing',
-        variant: 'accent' as const,
-        detail: pending.statusMessage || 'Transaction confirmed. Waiting for the operator to publish the workflow.',
-      };
-  }
-}
-
-function getWorkflowContractAddress(address: Address): Address | undefined {
-  return isContractDeployed(address) ? address : undefined;
-}
-
-function getWorkflowContractAddressForScope(
-  addrs: SandboxAddresses,
-  scope: WorkflowScope,
-): Address | undefined {
-  switch (scope) {
-    case 'sandbox':
-      return getWorkflowContractAddress(addrs.sandboxBlueprint);
-    case 'instance':
-      return getWorkflowContractAddress(addrs.instanceBlueprint);
-    case 'tee':
-      return getWorkflowContractAddress(addrs.teeInstanceBlueprint);
-  }
-}
-
-function getWorkflowIdentityKey(scope: WorkflowScope, workflowId: bigint | number) {
-  return `${scope}:${String(workflowId)}`;
-}
-
-function getWorkflowSortTimestamp(workflow: WorkflowRecord) {
-  if (workflow.kind === 'pending') {
-    return workflow.pending.createdAt;
-  }
-
-  return workflow.data.lastRunAt
-    ?? workflow.data.latestExecution?.executedAt
-    ?? workflow.data.nextRunAt
-    ?? 0;
-}
-
-function getOperatorLabel(scope: WorkflowScope) {
-  switch (scope) {
-    case 'sandbox':
-      return 'Sandbox operator';
-    case 'instance':
-      return 'Instance operator';
-    case 'tee':
-      return 'TEE operator';
-  }
-}
 
 export default function Workflows() {
   const queryClient = useQueryClient();
@@ -693,230 +579,4 @@ export default function Workflows() {
       </div>
     </ConsolePage>
   );
-}
-
-function WorkflowTable({
-  workflows,
-  onTrigger,
-  onCancel,
-  onResolvePending,
-  resolvingPendingKeys,
-  txPending,
-}: {
-  workflows: WorkflowRecord[];
-  onTrigger: (workflow: RemoteWorkflowRecord) => void;
-  onCancel: (workflow: RemoteWorkflowRecord) => void;
-  onResolvePending: (pending: PendingWorkflowCreation) => void;
-  resolvingPendingKeys: Record<string, boolean>;
-  txPending: boolean;
-}) {
-  return (
-    <div className="overflow-auto">
-      <table className="min-w-[980px] w-full table-fixed border-collapse">
-        <colgroup>
-          <col className="w-[25%]" />
-          <col className="w-[12%]" />
-          <col className="w-[12%]" />
-          <col className="w-[18%]" />
-          <col className="w-[13%]" />
-          <col className="w-[10%]" />
-          <col className="w-[10%]" />
-        </colgroup>
-        <thead>
-          <tr className="border-b border-[var(--sandbox-console-border)] bg-[var(--sandbox-console-surface)]">
-            {['Workflow', 'Status', 'Trigger', 'Target', 'Last run', 'Next run', 'Actions'].map((label) => (
-              <th
-                key={label}
-                className="px-3 py-2 text-left font-data text-[10px] font-semibold uppercase tracking-[0.12em] text-[var(--sandbox-console-muted)]"
-              >
-                {label}
-              </th>
-            ))}
-          </tr>
-        </thead>
-        <tbody>
-          {workflows.map((workflow) => (
-            <WorkflowTableRow
-              key={`${workflow.kind}:${workflow.blueprintId}:${String(workflow.id)}`}
-              workflow={workflow}
-              onTrigger={onTrigger}
-              onCancel={onCancel}
-              onResolvePending={onResolvePending}
-              pendingActionLoading={workflow.kind === 'pending' && !!resolvingPendingKeys[workflow.pending.key]}
-              txPending={txPending}
-            />
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
-}
-
-function WorkflowTableRow({
-  workflow,
-  onTrigger,
-  onCancel,
-  onResolvePending,
-  pendingActionLoading,
-  txPending,
-}: {
-  workflow: WorkflowRecord;
-  onTrigger: (workflow: RemoteWorkflowRecord) => void;
-  onCancel: (workflow: RemoteWorkflowRecord) => void;
-  onResolvePending: (pending: PendingWorkflowCreation) => void;
-  pendingActionLoading: boolean;
-  txPending: boolean;
-}) {
-  const triggerLabel: Record<string, string> = {
-    cron: 'Cron',
-    manual: 'Manual',
-  };
-
-  const isPending = workflow.kind === 'pending';
-  const name = isPending ? workflow.pending.name : workflow.data.name;
-  const triggerType = isPending ? workflow.pending.triggerType : workflow.data.triggerType;
-  const triggerConfig = isPending ? workflow.pending.triggerConfig : workflow.data.triggerConfig;
-  const status = isPending
-    ? getPendingWorkflowStatusPresentation(workflow.pending)
-    : getWorkflowStatusPresentation(workflow.data);
-  const detailPath = isPending ? null : buildWorkflowDetailPath(workflow.scope, workflow.id);
-  const canTrigger = !isPending && workflow.data.runnable && workflow.data.targetServiceId !== 0;
-  const canCancel = !isPending && workflow.data.active && workflow.data.targetServiceId !== 0;
-
-  return (
-    <tr className="group border-b border-[var(--sandbox-console-border)] transition-colors hover:bg-[var(--sandbox-console-surface)]">
-      <td className="px-3 py-3">
-        <div className="flex min-w-0 items-center gap-3">
-          <IdentityMark identity={getBlueprintIdentity(workflow.blueprintId)} size="md" />
-          <span className="min-w-0">
-            {detailPath ? (
-              <Link
-                to={detailPath}
-                className="block truncate font-display text-sm font-bold text-[var(--sandbox-console-text)] transition-colors hover:text-[var(--sandbox-console-brand)]"
-              >
-                {name || `Workflow #${String(workflow.id)}`}
-              </Link>
-            ) : (
-              <span className="block truncate font-display text-sm font-bold text-[var(--sandbox-console-text)]">
-                {name || `Workflow #${String(workflow.id)}`}
-              </span>
-            )}
-            <span className="block truncate font-data text-[11px] text-[var(--sandbox-console-subtle)]">
-              {workflow.kindLabel} · #{String(workflow.id)}
-            </span>
-          </span>
-        </div>
-      </td>
-      <td className="px-3 py-3">
-        <ConsoleChip tone={workflowStatusTone(status.variant)}>{status.label}</ConsoleChip>
-      </td>
-      <td className="px-3 py-3">
-        <span className="block truncate font-display text-sm font-bold text-[var(--sandbox-console-text)]">
-          {triggerLabel[triggerType] ?? triggerType}
-        </span>
-        <span className="block truncate font-data text-[11px] text-[var(--sandbox-console-subtle)]">
-          {triggerConfig || 'manual'}
-        </span>
-      </td>
-      <td className="px-3 py-3">
-        <span className="block truncate font-data text-xs font-bold text-[var(--sandbox-console-text)]">
-          {isPending
-            ? workflow.targetLabel
-            : workflow.data.targetStatus === 'missing'
-              ? `Target missing: ${workflow.targetLabel}`
-              : workflow.targetLabel}
-        </span>
-        <span className="block truncate font-data text-[11px] text-[var(--sandbox-console-subtle)]">
-          {status.detail}
-        </span>
-      </td>
-      <td className="px-3 py-3 font-data text-xs text-[var(--sandbox-console-muted)]">
-        {isPending ? formatWorkflowDate(workflow.pending.submittedAt, 'ms') : formatWorkflowDate(workflow.data.lastRunAt, 's')}
-      </td>
-      <td className="px-3 py-3 font-data text-xs text-[var(--sandbox-console-muted)]">
-        {isPending ? 'pending' : formatWorkflowDate(workflow.data.nextRunAt, 's')}
-      </td>
-      <td className="px-3 py-3">
-        <div className="flex flex-wrap items-center gap-2">
-          {isPending ? (
-            <WorkflowActionButton
-              tone="secondary"
-              onClick={() => onResolvePending(workflow.pending)}
-              disabled={pendingActionLoading}
-            >
-              {pendingActionLoading
-                ? 'Checking...'
-                : workflow.pending.status === 'awaiting-auth'
-                  ? 'Connect Operator'
-                  : 'Check Status'}
-            </WorkflowActionButton>
-          ) : null}
-          {!isPending && workflow.data.active && workflow.data.targetServiceId !== 0 ? (
-            <WorkflowActionButton
-              tone="success"
-              onClick={() => onTrigger(workflow)}
-              disabled={txPending || !canTrigger}
-              icon="i-ph:play-bold"
-            >
-              Trigger
-            </WorkflowActionButton>
-          ) : null}
-          {!isPending && canCancel ? (
-            <WorkflowActionButton
-              tone="secondary"
-              onClick={() => onCancel(workflow)}
-              disabled={txPending}
-              icon="i-ph:stop-bold"
-            >
-              Cancel
-            </WorkflowActionButton>
-          ) : null}
-        </div>
-      </td>
-    </tr>
-  );
-}
-
-function WorkflowActionButton({
-  children,
-  disabled,
-  icon,
-  onClick,
-  tone = 'secondary',
-}: {
-  children: ReactNode;
-  disabled?: boolean;
-  icon?: string;
-  onClick?: () => void;
-  tone?: 'secondary' | 'success';
-}) {
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      disabled={disabled}
-      className={cn(
-        'inline-flex h-8 items-center justify-center gap-1.5 rounded-[4px] border px-2.5 font-display text-xs font-bold transition-[background-color,border-color,box-shadow,color,transform] duration-150 active:scale-[0.98] disabled:cursor-not-allowed disabled:opacity-55',
-        tone === 'success'
-          ? 'border-[var(--sandbox-console-success-border)] bg-[var(--sandbox-console-success-soft)] text-[var(--sandbox-console-success)] hover:border-[var(--sandbox-console-success)] hover:bg-[color-mix(in_srgb,var(--sandbox-console-success)_18%,transparent)] hover:shadow-[inset_3px_0_0_var(--sandbox-console-success)]'
-          : 'border-[var(--sandbox-console-border)] bg-[var(--sandbox-console-control)] text-[var(--sandbox-console-secondary)] hover:border-[var(--sandbox-console-border-hover)] hover:bg-[var(--sandbox-console-control-hover)] hover:text-[var(--sandbox-console-text)] hover:shadow-[var(--sandbox-console-control-shadow-hover)]',
-      )}
-    >
-      {icon ? <span className={cn('text-xs', icon)} /> : null}
-      {children}
-    </button>
-  );
-}
-
-function workflowStatusTone(variant: 'stopped' | 'running' | 'secondary' | 'accent') {
-  if (variant === 'running') return 'ready';
-  if (variant === 'accent') return 'brand';
-  if (variant === 'stopped') return 'warn';
-  return 'muted';
-}
-
-function formatWorkflowDate(value: number | null | undefined, unit: 's' | 'ms') {
-  if (value == null || value <= 0) return '--';
-  const timestampMs = unit === 's' ? value * 1000 : value;
-  return new Date(timestampMs).toLocaleString();
 }
